% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flump.R
\name{flump}
\alias{flump}
\title{Dummy function}
\usage{
flump(x)
}
\arguments{
\item{x}{Any object or numeric or character or whatever}
}
\value{
The very object, that is provided to the function
}
\description{
Promt some shit and return the input provided
}
\examples{
temp1 <- flump(50);
temp2 <- flump( c(50, 63, 23) );
}
